Extract PotentialsDropdown into a resuable DropdownSelect
