+++ src/components/DropdownSelect/index.ts
@@ -1,0 +1,2 @@
+import DropdownSelect, { DropdownOption } from "./DropdownSelect";
+export { DropdownSelect, DropdownOption };
